fix comments & tests
