+++ apps/fyle/tasks.py
@@ -260,7 +260,7 @@ def import_and_export_expenses(report_id: str, org_id: str, is_state_change_even
         expense_group_ids = [expense_group['id'] for expense_group in expense_groups]
 
         if len(expense_group_ids) and not is_state_change_event:
-            export_to_xero(workspace.id, None, expense_group_ids, True, triggered_by=imported_from)
+            export_to_xero(workspace_id=workspace.id, export_mode=None, expense_group_ids=expense_group_ids, is_direct_export=True, triggered_by=imported_from)
 
     except WorkspaceGeneralSettings.DoesNotExist:
         logger.info('Configuration does not exist for workspace_id: %s', workspace.id)
@@ -270,57 +270,6 @@ def import_and_export_expenses(report_id: str, org_id: str, is_state_change_even
             task_log, _ = TaskLog.objects.update_or_create(workspace_id=workspace.id, type='FETCHING_EXPENSES', defaults={'status': 'IN_PROGRESS'})
 
         handle_import_exception(task_log)
-
-
-def post_accounting_export_summary(workspace_id: int, expense_ids: List = None, fund_source: str = None, is_failed: bool = False) -> None:
-    """
-    Post accounting export summary to Fyle
-    :param workspace_id: workspace id
-    :param expense_ids: expense ids
-    :param fund_source: fund source
-    :param is_failed: is failed
-    :return: None
-    """
-    # Iterate through all expenses which are not synced and post accounting export summary to Fyle in batches
-    fyle_credentials = FyleCredential.objects.get(workspace_id=workspace_id)
-    platform = PlatformConnector(fyle_credentials)
-    filters = {
-        'workspace_id': workspace_id,
-        'accounting_export_summary__synced': False
-    }
-
-    if expense_ids:
-        filters['id__in'] = expense_ids
-
-    if fund_source:
-        filters['fund_source'] = fund_source
-
-    if is_failed:
-        filters['accounting_export_summary__state'] = 'ERROR'
-
-    expenses_count = Expense.objects.filter(**filters).count()
-
-    accounting_export_summary_batches = []
-    page_size = 200
-    for offset in range(0, expenses_count, page_size):
-        limit = offset + page_size
-        paginated_expenses = Expense.objects.filter(**filters).order_by('id')[offset:limit]
-
-        payload = []
-
-        for expense in paginated_expenses:
-            accounting_export_summary = expense.accounting_export_summary
-            accounting_export_summary.pop('synced')
-            payload.append(expense.accounting_export_summary)
-
-        accounting_export_summary_batches.append(payload)
-
-    logger.info(
-        'Posting accounting export summary to Fyle workspace_id: %s, payload: %s',
-        workspace_id,
-        accounting_export_summary_batches
-    )
-    create_generator_and_post_in_batches(accounting_export_summary_batches, platform, workspace_id)
 
 
 def update_non_exported_expenses(data: Dict) -> None:
